(test): align existing tests with raw-eltype Constant contract
Tests that hard-coded the pre-refactor Float-widening assumptions, now updated
to the duck-type policy (Int in → Int out, Tg/Tv flow through unchanged):

- `test_grid_accessors.jl` — `_get_h(::Vector{Int}, …)` returns Int (not 1.0);
  `_get_inv_h` still Float via natural `inv(::Int) → Float64`.
- `test_complex_constant_series.jl` — Int range × Complex{Int} y stays
  `ConstantSeriesInterpolant{Int, Complex{Int}}`; Float32 grid stays Float32;
  scalar `sitp(0.5)` on Float32 keeps `Vector{Float32}` (no Tq-driven widening
  for plain numerical queries).
- `test_bc_complex_int.jl` — Float64 x + `Complex{Int}` y produces
  `ConstantInterpolant{Float64, Complex{Int}}` (Tv preserved).
- `test_type_stability.jl` — Int x with Constant produces
  `ConstantInterpolant{Int, Tv}`, not `{Float64}`.
- `test_promotion_alloc.jl` — `constant_interp(x_int, y_int, …) isa Vector{Int}`
  reflects duck-type (Linear/Quadratic/Cubic/etc. still `Vector{Float64}`).

+++ test/test_promotion_alloc.jl
@@ -119,9 +119,9 @@
         y_int = [0, 1, 3, 2, 5]
         xq_test = [0.5, 1.5, 2.5, 3.5]
 
-        # Allocating batch (all methods)
+        # Allocating batch (all methods). Constant duck-types: Int y → Vector{Int}.
         @test linear_interp(x_int, y_int, xq_test) isa Vector{Float64}
-        @test constant_interp(x_int, y_int, xq_test) isa Vector{Float64}
+        @test constant_interp(x_int, y_int, xq_test) isa Vector{Int}
         @test quadratic_interp(x_int, y_int, xq_test) isa Vector{Float64}
         @test cubic_interp(x_int, y_int, xq_test; extrap = ExtendExtrap()) isa Vector{Float64}
         @test pchip_interp(x_int, y_int, xq_test; extrap = ExtendExtrap()) isa Vector{Float64}
